refactor(super): split super ops service by domain
